Change dependencies order for Unix builds
In the Makefile, the `core-lib/.gitignore` rule was called to late, i.e,
after the building of `SOMCore.csp`. The building of the `SOMCore`
performs a `mv` that overrides the `Smalltalk` symlink.

By placing the `core-lib/.gitignore` as dependency of `CORE`, the git
sub-module is cloned before the building of the `SOMCore.csp` file, and
this time, the `mv` place the `.csp` in the `Smalltalk` symlink as it
should.